Remove set_sequence_number from WriteBatch API.
