+++ components/hybrid_engine/src/misc.rs
@@ -1,6 +1,6 @@
 // Copyright 2023 TiKV Project Authors. Licensed under Apache-2.0.
 
-use engine_traits::{KvEngine, MiscExt, RegionCacheEngine, Result};
+use engine_traits::{KvEngine, MiscExt, RegionCacheEngine, Result, WriteBatchExt};
 
 use crate::{engine::HybridEngine, hybrid_metrics::HybridEngineStatisticsReporter};
 
@@ -8,6 +8,7 @@ impl<EK, EC> MiscExt for HybridEngine<EK, EC>
 where
     EK: KvEngine,
     EC: RegionCacheEngine,
+    HybridEngine<EK, EC>: WriteBatchExt,
 {
     type StatisticsReporter = HybridEngineStatisticsReporter;
